+++ src/Module/Platform/User/Dto/UserInviteInputFactory.php
@@ -17,6 +17,7 @@ class UserInviteInputFactory implements UserInviteInputFactoryInterface
             email: Str::trimFromArray($data, 'email'),
             role: Str::trimFromArray($data, 'role'),
             message: Str::trimOrNullFromArray($data, 'message'),
+            disabled: (bool) ($data['disabled'] ?? false),
         );
     }
 }